// frontend/src/api/generated/zod/order/order.zod.ts
export const draftOrderBodyCustomerOrderCodeMin = 0;
export const draftOrderBodyCustomerOrderCodeMax = 30;

export const draftOrderBodyLinesItemQuantityMax = 9999;

export const draftOrderBodyLinesMin = 0;
export const draftOrderBodyLinesMax = 100;



export const DraftOrderBody = zod.object({
  "customerOrderCode": zod.string().min(draftOrderBodyCustomerOrderCodeMin).max(draftOrderBodyCustomerOrderCodeMax).describe('客先注文番号。注文ごとに一意にする'),
  "lines": zod.array(zod.object({
  "productId": zod.uuid().describe('注文する商品の ID'),
  "quantity": zod.int().min(1).max(draftOrderBodyLinesItemQuantityMax).optional().describe('注文する数量')
}).describe('注文する商品と数量。')).min(draftOrderBodyLinesMin).max(draftOrderBodyLinesMax).describe('注文の明細')
}).describe('下書きの注文を作る API の本文。')
// ...
export const changeOrderLinesBodyLinesItemQuantityMax = 9999;

export const changeOrderLinesBodyLinesMin = 0;
export const changeOrderLinesBodyLinesMax = 100;




export const ChangeOrderLinesBody = zod.object({
  "lines": zod.array(zod.object({
  "productId": zod.uuid().describe('注文する商品の ID'),
  "quantity": zod.int().min(1).max(changeOrderLinesBodyLinesItemQuantityMax).optional().describe('注文する数量')
}).describe('注文する商品と数量。')).min(changeOrderLinesBodyLinesMin).max(changeOrderLinesBodyLinesMax).describe('置き換えた後の明細'),
  "lockNo": zod.int().min(1).optional().describe('画面が読んだ注文のロック番号')
}).describe('下書きの注文の明細を置き換える API の本文。')